Maxwell Studio/requirements.txt
pyside6
pyqtgraph
numpy
pyserial